Can successfully run Webpack to create dist/bundle.js
The bundled output is noticeably different; likely due to Webpack 4
vs. 2/3. The order of dependencies also appears to have changed.
The overall output *seems* to be the same. Unsure how to full test
that everything is getting bundled appropriately.

This also allows the project to be built without having to manually
copy files into node_modules.