[refactor, chore] Including TSLint; fixing most things.
As it currently stands, bringing in TSLint with *all* rules enabled by
default and override each as we go. This means that as we upgrade and
bring in new rules we might start seeing new rules that the code
violates.

Because all the rules are enabled by default, the choice to explicitly
disable/relax rules can be discussed as part of PRs.

This also brought in strict type checking. This means we shouldn't be
relying on truthy/falsey value coercion. It also means that we will
likely need to create some type definition files or port certain files
over. That...will be a bit tedious, but will make the codebase better as
a whole.