[chore] Disable "no-inferrable-types" lint rule
After discussing the conflicting 'typedef' vs. 'no-inferrable-types'
TSLint rules (https://github.com/palantir/tslint/issues/711) in the
Discord #development channel, the conclusion was to disable this rule in
order to be uniform and explicit in what all declarations are. Although
it's more verbose, it will help us out in the long run.