[refactor] Determine program execution from terminal via hashmap instead of switch statement.
Introducing variables within the case block of a switch statement can be
problematic and confusing as case statements do not have their own
scope, and thus are hoisted to the function level. Thus, with a large
enough switch statement, variable declarations could easily be
overlooked as they are peppered throughout the cases. This is especially
problematic for 'const' and 'let' declarations, which are very concerned
about scoping. There is an ESLint rule specifically enabled to catch
this behavior.

This refactoring is a way to mitigate the behavior, while still allowing
for the use of 'const'.