Update AugmentationHelpers.jsx
Extensive rewording of augmentation to descriptions. The purpose of this was to condense sentence structure, remove duplication, and create more robust and interesting descriptions of iterated augmentations. The descriptions should read from a third person perspective, similar to if the player were reading a manual with information about the augment.