Blackjack pays 1.5x, fix appearances
I'll lead with this so it doesn't look like I snuck it in: Winning by blackjack is special and usually pays 1.5x so I made it so :) All the infrastructure was already in place.

Fixed bkack-on-black text, I didn't even know the game kept count of your cards for you! Improved the way it's displayed a bit (screenshot in PR)

Also simplified the code that determines gains and applies the win since there was some duplication there.