Improve the readability of text in the theme editor.
Copy the string above if you want to backup or share your theme with others. -> Backup your theme or share it with others by copying the string above.
Use the buttons below to replace the current theme with a pre-built template -> Replace the current theme with a pre-built template using the buttons below.