feat(command:buy): show owned items in buy list
Just a little output change when you `buy -l`, to show you which items
you already own.
I couldn't access the theme when importing it here for use, it kept
crashing the app so the [OWNED} shows up always as green.. Which isn't
the end of the world considering the money always shows up in Yellow.
I think this area of code/output could do with a little bit of love at
some point.

Resolves danielyxie/bitburner#2069