Added wrapper system for Netscript1
The method startNetscript1 will now look for methods on the NS object that have a specific prefix that indicates they are intended to replace Netscript 2 equivalent methods. If such a method exists, it will ignore the method without the prefix and use the one with the prefix, stripping off the prefix.
A helper method called addWrapper does the work of creating an NS1 method that wraps an NS2 method and makes its return value into something NS1 can understand.