Apply bitnode multiplier before capping at 1
It is entirely possible to hack more than the BN multiplier, which means this function is reporting the effects of the multiplier incorrectly. By moving the BN multiplier before to before the caps are applied, the functionality is corrected.

(ie, hack % * BN multiplier, THEN cap to 1; not cap to 1, then BM multiply the 1)

TESTED: I'm currently running in BN 5 ( * BitNodeMultipliers.ScriptHackMoney = 0.15), and sent a 4b thread hack against several fully grown servers (Avail Money = 100% of Max Money). Each went down to 0% (not 85% as the current implementation would suggest) after the hack hit