Hacknet server achievements grant associated hacknet node achieve
fixes #1905

Added a new property to the Achievement interface to contain a list
of additional achievements that should be granted when an achievement's
conditional passes.

Right now this is only used to unlock the 4 hacknet node achieves
when the corresponding hacknet server achievement is awarded.

This branch also changes the calculateAchievements function so it
only runs the conditionals for achievements the player doesn't have,
instead of checking all of them and then filtering out the unowned.