Change job shortcut test to match sidebar item
I switched it to match the same condition as the sidebar menu item,
which checks for player.companyName, which should be set to the latest
company the player worked at. This can prevent a crash loading the jobs
page with an invalid companyName.

This fixes a problem a user had where the player.companyName was empty
while actually items in player.jobs, but I could not figure out how
to replicate. It may have had something to do with part-time jobs.