Add Electron preload script to allow communication
Adds a channel to communicate between the main process & the renderer
process, so that the game can easily ship data back to the main process.

It uses the Electron contextBridge & ipcRenderer/ipcMain.

Connects those events to various save functions. Adds triggered events
on game save, game load, and imported game. Adds way for the Electron
app to ask for certain actions or data.

This could be used to replace for achievements, toasts, etc.

- Trigger Save
- Trigger Game Export
- Trigger Scripts Export
- Push Game Saved
- Push Game Ready
- Push Import Result
- Request Save Data
- Request Save Info