0/0 catches
Realized that it's not necessary to set oldMoney to 0 in the case of a zero max money server (a later check takes care of that), So allowing oldMoney to be at least 1 solves a few problems of the divide by 0 sort. But went ahead and protected those /0 calls as well.