Add 0xDEADBEEF exploit (SF-1)
Adds a new exploit that requires the player to bruteforce a passphrase
(or read the source to reverse the seed method).

Adds a new `unlock` command that adds a new server to the network once a
passphrase was successfully matched. This needs to be done from the
terminal, no direct ns script access.

Once created, the server is added as a hidden server which hides it from
scan, scan-analyze, autocomplete, ns.scan. This server is still able to be
connected to given the proper route.

Since the hostname is hex, it trips the connect method and
requires the player to wrap its hostname in quotes.
Once that server is backdoored, the exploit will be awarded and
the achievement unlocked.