Fixed a bug in a stock market test that caused false negatives
The test was incorrectly creating a reference to a live stock
object instead of making a copy. This resulted in all changes to
the live version being reflected in what was supposed to be a
snapshot of a past state, and thus a false negative test result.
I have fixed this by creating a shallow copy of the object, but
am certain cleaner solutions exist.