CCT: inconsistent probability for generation between online and offline
Fixes #4110.  While online, there is 25% chance for a Coding Contract to appear.  The same should apply when a player loads the game after being offline for a period of time.  Currently, loading the game after being offline would give a generation probability of less than 25%.  The mismatch between offline and online gives an unfair advantage to having the game running for an extended period of time.