CCT: no special characters in file names of Coding Contracts
Fixes #4067.  The file name of a Coding Contract follows the format contract-xxx-rewardName.cct.  The part `xxx` means a sequence of random decimal digits.  The part `rewardName` can be an empty string if the player is not part of any faction nor is working for a company.  However, if the player is working for a company or faction whose name has a special character, then the special character would also appear in the generated file name.  We only want alphanumeric characters throughout the whole file name.